+++ docs/index.jsx
@@ -21,8 +21,6 @@ import Checkbox from '../src/components/Checkbox';
 import CheckboxExample from '../src/components/Checkbox/example';
 import Code from '../src/components/Code';
 import CodeExample from '../src/components/Code/example';
-import IconSample from '../src/components/IconSample';
-import IconExample from '../src/components/IconSample/example';
 import Input from '../src/components/Input';
 import InputExample from '../src/components/Input/example';
 import Label from '../src/components/Label';
@@ -77,10 +75,6 @@ const componentsArray = [
   {
     index: Input,
     example: InputExample,
-  },
-  {
-    index: IconSample,
-    example: IconExample,
   },
   {
     index: Label,