+++ docs/ComponentRow.js
@@ -38,6 +38,7 @@ const ComponentRow = (props) => {
       <Code
         type="block"
         hasCopyButton={ true }
+        isHighlighted={ true }
         language="html">
         { codeHTML }
       </Code>