+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+
+## [22.2.0-rc.1][22.2.0-rc.1] - 2016-12-06
 ### Added
 - [Feature] Bring back syntax highlighting for Code component and load asynchronously. (#678)
 
@@ -710,7 +712,7 @@ _This version was accidentally released and contains no changes._
 ### Added
 - [Patch] Use semantic versioning in LEGO (#58).
 
-[Unreleased]: https://github.com/optimizely/oui/compare/v22.1.0...HEAD
+[Unreleased]: https://github.com/optimizely/oui/compare/v22.2.0-rc.1...HEAD
 [0.0.2]: https://github.com/optimizely/oui/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/optimizely/oui/compare/v0.0.2...v0.0.3
 [1.0.0]: https://github.com/optimizely/oui/compare/v0.0.3...v1.0.0
@@ -782,3 +784,4 @@ _This version was accidentally released and contains no changes._
 [21.0.0]: https://github.com/optimizely/oui/compare/v20.0.0...v21.0.0
 [22.0.0]: https://github.com/optimizely/oui/compare/v21.0.0...v22.0.0
 [22.1.0]: https://github.com/optimizely/oui/compare/v22.0.0...v22.1.0
+[22.2.0-rc.1]: https://github.com/optimizely/oui/compare/v22.1.0...v22.2.0-rc.1